feat(engine): fetching and completion of external tasks
- an external task can be defined for a serviceTask in the BPMN XML
  by setting the camunda:type attribute to 'external' and specifying
  a camunda:topic attribute that identifies the task's topic
- adds a new service: ExternalTaskService
- via ExternalTaskService, a number of tasks can be fetched for
  a specific topics along with variables for each return external
  task instance
- returned tasks are locked for a user-defined time during which they
  cannot be locked/completed by other workers
- via ExternalTaskService, tasks can be completed
- task fetching does not fail on OptimisticLockingExceptions for
  individual tasks

related to CAM-4619

+++ engine/src/main/java/org/camunda/bpm/engine/impl/cmmn/entity/repository/CaseDefinitionQueryImpl.java
@@ -111,7 +111,7 @@ public class CaseDefinitionQueryImpl extends AbstractQuery<CaseDefinitionQuery, 
 
   public CaseDefinitionQuery caseDefinitionVersion(Integer caseDefinitionVersion) {
     ensureNotNull(NotValidException.class, "version", caseDefinitionVersion);
-    ensurePositive(NotValidException.class, "version", caseDefinitionVersion);
+    ensurePositive(NotValidException.class, "version", caseDefinitionVersion.longValue());
     this.version = caseDefinitionVersion;
     return this;
   }
